+++ packages/sui-critical-css/src/extract-from-url.js
@@ -59,7 +59,7 @@ export async function extractCSSFromUrl({
     await page.coverage.startCSSCoverage()
 
     const response = await page
-      .goto(url, {waitUntil: 'load'})
+      .goto(url, {waitUntil: 'networkidle'})
       .catch(error => ({error}))
 
     const closeAll = async error => {
@@ -71,8 +71,7 @@ export async function extractCSSFromUrl({
 
     if (!response) await closeAll('Response is not present')
 
-    if (response.error)
-      await closeAll(`Response has an error: ${response.error}`)
+    if (response.error) await closeAll(response.error)
 
     if (!response.ok() && response.status() !== 304)
       await closeAll(`Response status code ${response.status()} for url ${url}`)