+++ packages/sui-react-router/src/match.js
@@ -1,3 +1,4 @@
+// @ts-check
 // from: https://github.com/ReactTraining/react-router/blob/v3/modules/match.js
 
 import {REPLACE} from 'history/lib/Actions'