+++ packages/sui-react-router/src/withRouter.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 import {useRouter} from './hooks'
 import hoistNonReactStatics from 'hoist-non-react-statics'