+++ packages/sui-ssr/server/ssr/index.js
@@ -168,7 +168,20 @@ export default async (req, res, next) => {
 
   const {__HTTP__} = initialProps
   if (__HTTP__) {
-    const {redirectTo, redirectStatusCode} = __HTTP__
+    const {redirectTo, redirectStatusCode, httpCookie, headers} = __HTTP__
+
+    if (httpCookie) {
+      const [[field, value]] = Object.entries(httpCookie)
+      res.cookie(field, value, {httpOnly: true})
+    }
+
+    if (headers) {
+      headers.forEach(header => {
+        const [[field, value]] = Object.entries(header)
+        res.append(field, value)
+      })
+    }
+
     if (redirectTo) {
       const isValidRedirectStatusCode = redirectStatusCodes.includes(
         redirectStatusCode