+++ packages/sui-studio/src/runtime-mocha/index.js
@@ -8,8 +8,20 @@ addSetupEnvironment(window)
 window.__STUDIO_CONTEXTS__ = {}
 window.__STUDIO_COMPONENT__ = {}
 
+const pattern = process.env.CATEGORIES
+const categories = pattern ? pattern.split(',') : null
+
+const filterAll = key => {
+  const [, category] = key.split('/')
+  return !categories || categories.includes(category)
+}
+
 // Require all the files from a context
-const importAll = request => request.keys().forEach(request)
+const importAll = request =>
+  request
+    .keys()
+    .filter(filterAll)
+    .forEach(request)
 
 // Avoid running Karma until all components tests are loaded
 const originalKarmaLoader = window.__karma__.loaded
@@ -23,44 +35,52 @@ const testsFiles = require.context(
 
 // get all the needed components from the available tests
 Promise.all(
-  testsFiles.keys().map(async key => {
-    // get the category component from the segments of the path
-    // ex: ./card/property/index.js -> card property
-    const [, category, component] = key.split('/')
-    const categoryComponentKey = `${category}/${component}`
+  testsFiles
+    .keys()
+    .filter(filterAll)
+    .map(async key => {
+      // get the category component from the segments of the path
+      // ex: ./card/property/index.js -> card property
+      const [, category, component] = key.split('/')
+      const categoryComponentKey = `${category}/${component}`
 
-    const getContexts = await importContexts({category, component})
-    const contexts =
-      typeof getContexts === 'function' ? await getContexts() : getContexts
+      const getContexts = await importContexts({category, component})
+      const contexts =
+        typeof getContexts === 'function' ? await getContexts() : getContexts
 
-    const componentModule = await importReactComponent({
-      category,
-      component,
-      extractDefault: true
+      const componentModule = await importReactComponent({
+        category,
+        component,
+        extractDefault: true
+      })
+
+      const Component =
+        componentModule && (componentModule.type || componentModule)
+
+      if (!componentModule) {
+        console.error(
+          `Could not find component ${categoryComponentKey} in ${key}`
+        )
+        console.error(`Available keys: `, testsFiles.keys())
+        throw new Error(`Missing export default from ${categoryComponentKey}`)
+      }
+
+      const {displayName} = Component
+      // store on the window the contexts and components using the ${category/component} key
+      window.__STUDIO_CONTEXTS__[categoryComponentKey] = contexts
+      window.__STUDIO_COMPONENT__[categoryComponentKey] = Component
+
+      const ComponentWithReactContext = addReactContextToComponent(Component, {
+        context: contexts.default,
+        id: categoryComponentKey
+      })
+
+      // if the Component has a displayName, we're going to make the Component
+      // available through the window object with the default context without importing it
+      if (displayName) {
+        window[displayName] = ComponentWithReactContext
+      }
     })
-
-    const Component =
-      componentModule && (componentModule.type || componentModule)
-    if (!componentModule) {
-      throw new Error(`Missing export default from ${categoryComponentKey}`)
-    }
-
-    const {displayName} = Component
-    // store on the window the contexts and components using the ${category/component} key
-    window.__STUDIO_CONTEXTS__[categoryComponentKey] = contexts
-    window.__STUDIO_COMPONENT__[categoryComponentKey] = Component
-
-    const ComponentWithReactContext = addReactContextToComponent(Component, {
-      context: contexts.default,
-      id: categoryComponentKey
-    })
-
-    // if the Component has a displayName, we're going to make the Component
-    // available through the window object with the default context without importing it
-    if (displayName) {
-      window[displayName] = ComponentWithReactContext
-    }
-  })
 )
   .then(() => {
     // in order to force all tests, we're importing all the files that matches the pattern