+++ packages/babel-preset-sui/CHANGELOG.md
@@ -1,5 +1,68 @@
 # CHANGELOG
 
+# 3.32.0 (2021-11-24)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Release package without changelog ([a67f0cb](https://github.com/SUI-Components/sui/commit/a67f0cb5d655d5e215c5f2f6ac1266badf721c35))
+
+
+
+# 3.31.0 (2021-11-17)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Upgrade babel dependencies ([a3a82e2](https://github.com/SUI-Components/sui/commit/a3a82e294627153773c4bca14343b545fdf5e943))
+
+
+
+# 3.30.0 (2021-11-05)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Upgrade dependencies ([044d4d9](https://github.com/SUI-Components/sui/commit/044d4d99befa98bf47abc077bded050a9639b485))
+
+
+
+# 3.29.0 (2021-10-21)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Upgrade dependencies about babel ([d59044e](https://github.com/SUI-Components/sui/commit/d59044e41ba2f6fc861b3b5ed01c156f70bfc904))
+
+
+
+# 3.28.0 (2021-09-28)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Upgrade dependencies ([d0ce85c](https://github.com/SUI-Components/sui/commit/d0ce85cfad0b09b57ac562cd5b351ae1487c9f6b))
+
+
+
+# 3.27.0 (2021-09-08)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Upgrade deps for babel-preset-sui ([ac95437](https://github.com/SUI-Components/sui/commit/ac9543765f5f01dfaa3e2556c76fb0a105b352fd))
+
+
+
+# 3.26.0 (2021-08-16)
+
+
+### Features
+
+* **packages/babel-preset-sui:** Use latest babel versions ([0f30814](https://github.com/SUI-Components/sui/commit/0f308141902e8f683c91913c870cd3617b5f6545))
+
+
+
 # 3.25.0 (2021-06-22)